+++ kriegspiel/__init__.py
@@ -4,7 +4,7 @@ __author__ = "Alexander Filatov"
 
 __email__ = "alexander@kriegspiel.org"
 
-__version__ = "1.3.2"
+__version__ = "1.3.3"
 
 from kriegspiel.berkeley import BerkeleyGame
 from kriegspiel.game import KriegspielGame
@@ -14,16 +14,22 @@ from kriegspiel.move import KriegspielMove
 from kriegspiel.move import MainAnnouncement
 from kriegspiel.move import QuestionAnnouncement
 from kriegspiel.move import SpecialCaseAnnouncement
+from kriegspiel.snapshot import BerkeleyGameSnapshot
+from kriegspiel.snapshot import KriegspielGameSnapshot
+from kriegspiel.snapshot import ScoresheetSnapshot
 from kriegspiel.wild16 import Wild16Game
 
 __all__ = [
     "BerkeleyGame",
+    "BerkeleyGameSnapshot",
     "CapturedPieceAnnouncement",
     "KriegspielAnswer",
     "KriegspielGame",
+    "KriegspielGameSnapshot",
     "KriegspielMove",
     "MainAnnouncement",
     "QuestionAnnouncement",
+    "ScoresheetSnapshot",
     "SpecialCaseAnnouncement",
     "Wild16Game",
 ]